Changed everything
Divide an conquer;
Added a ton of comments
Added custom props for the character colors
Wrote the init funcion with an onload trigger, changed the character colors array into an object, had to rewrite logic from array based to object based (for ... in loop).
Added a custom properties object
Store the objects into localftorage
Added a function to write changes to the localstorage.